requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
fyers-apiv3==3.1.3
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
APScheduler==3.10.4
SQLAlchemy==2.0.30
requests==2.32.3
pytz==2024.1
Werkzeug==3.0.3
WTForms==3.1.2
email-validator==2.2.0
gunicorn==22.0.0
